@@ -14,7 +14,7 @@ Console.WriteLine("Agora digite o preço por hora:");
 precoPorHora = Convert.ToDecimal(Console.ReadLine());
 
 // Instancia a classe Estacionamento, já com os valores obtidos anteriormente
-Estacionamento es = new Estacionamento(precoInicial, precoPorHora);
+Estacionamento es = new(precoInicial, precoPorHora);
 
 string opcao = string.Empty;
 bool exibirMenu = true;
@@ -29,27 +29,36 @@ while (exibirMenu)
     Console.WriteLine("3 - Listar veículos");
     Console.WriteLine("4 - Encerrar");
 
-    switch (Console.ReadLine())
+    try
     {
-        case "1":
-            es.AdicionarVeiculo();
-            break;
+        switch (Console.ReadLine())
+        {
+            case "1":
+                es.AdicionarVeiculo();
+                break;
 
-        case "2":
-            es.RemoverVeiculo();
-            break;
+            case "2":
+                es.RemoverVeiculo();
+                break;
 
-        case "3":
-            es.ListarVeiculos();
-            break;
+            case "3":
+                es.ListarVeiculos();
+                break;
 
-        case "4":
-            exibirMenu = false;
-            break;
+            case "4":
+                exibirMenu = false;
+                break;
 
-        default:
-            Console.WriteLine("Opção inválida");
-            break;
+            default:
+                Console.WriteLine("Opção inválida");
+                break;
+        }
+
+    }
+    catch (Exception ex)
+    {
+        Console.WriteLine(ex.Message);
+        Console.WriteLine();
     }
 
     Console.WriteLine("Pressione uma tecla para continuar");
